fix: simplify filters, conditions, move schemas to separate folders
