+++ packages/apps/human-app/frontend/src/modules/signup/operator/components/add-keys/existing-keys-form.tsx
@@ -9,7 +9,7 @@ import { useEditExistingKeysMutation } from '../../hooks';
 import {
   type EditEthKVStoreValuesMutationData,
   getEditEthKVStoreValuesMutationSchema,
-} from '../../utils';
+} from '../../schema';
 import { EditExistingKeysForm } from './edit-existing-keys-form';
 import { ExistingKeys } from './existing-keys';
