+++ packages/apps/human-app/frontend/src/modules/signup/operator/components/add-stake/stake-form.tsx
@@ -15,7 +15,7 @@ import { useAddStake } from '../../hooks';
 import {
   type AddStakeCallArguments,
   addStakeAmountCallArgumentsSchema,
-} from '../../utils';
+} from '../../schema';
 
 export function StakeForm({
   decimals,
