+++ packages/apps/human-app/frontend/src/modules/signup/operator/hooks/use-add-stake-mutation-state.ts
@@ -3,7 +3,7 @@ import { useMutationState } from '@tanstack/react-query';
 import type { MutationState } from '@tanstack/react-query';
 import { useConnectedWallet } from '@/shared/contexts/wallet-connect';
 import type { ResponseError } from '@/shared/types/global.type';
-import { type AddStakeCallArguments } from '../utils';
+import { type AddStakeCallArguments } from '../schema';
 
 export function useAddStakeMutationState() {
   const { address } = useConnectedWallet();
