+++ packages/apps/human-app/frontend/src/modules/signup/operator/schema/index.ts
@@ -1,0 +1,2 @@
+export * from './add-stake-amount-call-arguments-schema';
+export * from './eth-kv-store-values-mutation-schema';
